1.2
model层优化及修改

+++ Addons/V/Model/Bloodpress.php
@@ -13,26 +13,36 @@ class Bloodpress
 {
     /**
      * 按时间戳删除血压记录
-     * @param $time
+     * @param int $time
+     * @param int $userId
      * @return boolean
      */
-    public function deleteBloodLogByTimestamp($time)
+    public function deleteBloodLogByTimestamp($time,$userId=null)
     {
+        //参数校验
         $time = intval($time);
-        $userId = bus('tokenInfo')['userId'];
+        $userId = $userId?:bus('tokenInfo')['userId'];
+        $userId = intval($userId);
+        if(empty($time)||empty($userId))return false;
+        //执行sql
         $check = server('Db')->query("delete from bloodpress where `userId`= $userId and `time`= $time");
         return $check?true:false;
     }
 
     /**
      * 按日期删除血压记录
-     * @param $createDay
+     * @param int $createDay
+     * @param int $userId
      * @return boolean
      */
-    public function deleteBloodLogByDate($createDay)
+    public function deleteBloodLogByDate($createDay,$userId=null)
     {
+        //参数校验
         $createDay = intval($createDay);
-        $userId = bus('tokenInfo')['userId'];
+        $userId = $userId?:bus('tokenInfo')['userId'];
+        $userId = intval($userId);
+        if(empty($createDay)||empty($userId)) return  false;
+        //执行sql
         $check = server('Db')->query("delete from bloodpress where `userId`= $userId and `createDay`= $createDay");
         return $check?true:false;
     }
@@ -40,50 +50,58 @@ class Bloodpress
     /**
      * 插入血压记录
      * @param array $req
+     * 请求参数 包含：type,createDay,time,shrink,diastole,bpm,day,userId
      * @return boolean
      */
     public function insertBloodLog(Array $req)
     {
-        //插入bloodpress表
-        $userId = bus('tokenInfo')['userId'];
-        $insert = array();
-        foreach($req['story'] as $v){
-            $v['userId'] = $userId;
-            $insert[] = '('.implode(',',$v).')';
-        }
-        $insert = implode(',',$insert);
-        $check = server('Db')->query("insert into bloodpress(type,createDay,time,shrink,diastole,bpm,day,userId)values $insert");
-        return $check?true:false;
+        //校验参数的存在性和非空
+        $fields = ['type','createDay','time','shrink','diastole','bpm','day','userId'];   //校验标准
+        $userId = $req['userId']?:bus('tokenInfo')['userId'];
+        $req['userId'] = intval($userId);
+        if(!model('Validate')->validateParams($fields,$req)) return false;
+        //执行sql
+        $insert = server('Db')->autoExecute('user', $req, 'INSERT');
+        return $insert?true:false;
     }
 
     /**
      * 通过日期和类型获取血压记录
      * @param array $req
+     * $req 包含键名有：type,createDay,userId
      * @return array
      */
     public function getBloodLogByDateAndType(Array $req)
     {
+        //校验参数
         $userId = $req['userId']?:bus('tokenInfo')['userId'];
+        $userId = intval($userId);
         $type = intval($req['type']);
         $createDay = intval($req['createDay']);
+        if(empty($userId)||empty($type)||empty($createDay)) return [];
+        //执行sql
         $bloodInfo = server('Db')->getAll("select time,shrink,diastole,bpm,createDay from bloodpress where `userId`=$userId and `type`=$type and `createDay`=$createDay");
         return $bloodInfo?$bloodInfo:[];
     }
 
     /**
      * 通过日期获取血压折线图
-     * @param $createDay
-     * @param $day
+     * @param int $createDay
+     * @param int $day
      * 1:白天，0：晚上，折线图未用到，方便饼状图使用
+     * @param int $userId
      * @return array
      */
     public function getBloodLineGraphByDate($createDay,$day=null,$userId=null)
     {
+        //参数处理
         $userId = $userId?:bus('tokenInfo')['userId'];
         $createDay = intval($createDay);
-        //$peiChart是饼状图所查询条件，此处并未使用
-        $peiChart = $day||$day===0?"and `day`=$day":'';
+        if(empty($userId)||empty($createDay)||!is_int($day)) return [];
+        $peiChart = $day||$day===0?"and `day`='$day'":'';
+        //执行sql
         $bloodInfo = server('Db')->getAll("select shrink,diastole,bpm from bloodpress where `userId`=$userId and `createDay`='$createDay' $peiChart");
+        //返回值处理
         if($bloodInfo){
             $data = array();
             foreach($bloodInfo as $v){
@@ -98,9 +116,10 @@ class Bloodpress
 
     /**
      * 通过日期获取血压柱状图
-     * @param $createDay
-     * @param $day
+     * @param int $createDay
+     * @param int $day
      * 1:白天，0：晚上，柱状图未用到，方便饼状图使用
+     * @param int $userId
      * @return array
      */
     public function getBloodBarGraphByDate($createDay,$day=null,$userId=null)
@@ -137,8 +156,9 @@ class Bloodpress
 
     /**
      * 根据白天夜晚或全天获取饼状图
-     * @param $createDay
-     * @param $day
+     * @param int $createDay
+     * @param string $day
+     * $day可选值：day,night,'',空代表全天
      * @return array
      */
     public function getPieChartByDay($createDay,$day='')
@@ -169,7 +189,7 @@ class Bloodpress
             $countValue=array_count_values($allValue);   //对数据进行统计出现次数
             $numValue = array();   //初始化存储次数的数组
             foreach($countValue as $k=>$v){
-                $key = floor($k/10);     //数据分割以10为单位
+                $key = intval(floor($k/10));     //数据分割以10为单位
                 $numValue[$key][] =$v;
             }  //将十位以上相同的数据分组
             if(!empty($numValue)){
